合同用印文件预览word pdf

+++ blade-service-api/blade-contract-api/src/main/java/org/springblade/contract/entity/CglProofingContract1Entity.java
@@ -52,7 +52,7 @@ public class CglProofingContract1Entity extends BaseEntity {
 	 * 数量
 	 */
 	@ApiModelProperty(value = "数量")
-	private Long number;
+	private Double number;
 	/**
 	 * 合计金额（元，含税）
 	 */
